fix: redurant api
